Issue #113 Move notifications types to their own file
